+++ src/lib/init.ts
@@ -35,6 +35,8 @@ import {
   getMainAquaFileContent,
   READMEs,
   JS_CLIENT_NPM_DEPENDENCY,
+  README_MD_FILE_NAME,
+  TYPESCRIPT_NPM_DEPENDENCY,
 } from "../lib/const.js";
 import {
   ensureFrontendCompiledAquaPath,
@@ -45,12 +47,18 @@ import {
   setProjectRootDir,
   getREADMEPath,
   projectRootDir,
-  getPackageJSONPath,
+  getFrontendPackageJSONPath,
   getFrontendSrcPath,
   getFrontendCompiledAquaPath,
   getIndexHTMLPath,
-  getTsConfigPath,
+  getFrontendTsConfigPath,
   getFrontendPath,
+  ensureGatewayCompiledAquaPath,
+  getGatewayServerTSorJSPath,
+  getGatewayPackageJSONPath,
+  getGatewaySrcPath,
+  getGatewayTsConfigPath,
+  getGatewayPath,
 } from "../lib/paths.js";
 import { confirm, input, list } from "../lib/prompt.js";
 import CLIPackageJSON from "../versions/cli.package.json" assert { type: "json" };
@@ -197,24 +205,22 @@ export async function init(options: InitArg = {}): Promise<FluenceConfig> {
 
     case "quickstart": {
       await quickstart();
+      await initTSorJSGatewayProject({ isJS: true, fluenceConfig });
       break;
     }
 
     case "js": {
       await quickstart();
       await initTSorJSProject({ isJS: true, fluenceConfig });
+      await initTSorJSGatewayProject({ isJS: true, fluenceConfig });
       break;
     }
 
     case "ts": {
       await quickstart();
       await initTSorJSProject({ isJS: false, fluenceConfig });
+      await initTSorJSGatewayProject({ isJS: false, fluenceConfig });
       break;
-    }
-
-    default: {
-      const _exhaustiveCheck: never = template;
-      return _exhaustiveCheck;
     }
   }
 
@@ -254,45 +260,18 @@ export async function init(options: InitArg = {}): Promise<FluenceConfig> {
 }
 
 const shouldInit = async (projectPath: string): Promise<boolean> => {
-  if (!isInteractive) {
-    return true;
-  }
+  let directoryContent: string[];
 
-  const pathDoesNotExists = !existsSync(projectPath);
-
-  if (pathDoesNotExists) {
-    return true;
-  }
-
-  const pathIsNotADirectory = !(await stat(projectPath)).isDirectory();
-
-  if (pathIsNotADirectory) {
-    return true;
-  }
-
-  const directoryContent = await readdir(projectPath);
-  const pathIsEmptyDir = directoryContent.length === 0;
-
-  if (pathIsEmptyDir) {
-    return true;
-  }
-
-  const dirHasOnlyGitInside =
-    directoryContent.length === 1 && directoryContent[0] === ".git";
-
-  if (dirHasOnlyGitInside) {
-    return true;
-  }
-
-  const hasUserConfirmedInitInNonEmptyDir = await confirm({
-    message: `Directory ${color.yellow(projectPath)} is not empty. Proceed?`,
-  });
-
-  if (hasUserConfirmedInitInNonEmptyDir) {
-    return true;
-  }
-
-  return false;
+  return (
+    !isInteractive ||
+    !existsSync(projectPath) ||
+    !(await stat(projectPath)).isDirectory() ||
+    (directoryContent = await readdir(projectPath)).length === 0 ||
+    (directoryContent.length === 1 && directoryContent[0] === ".git") ||
+    (await confirm({
+      message: `Directory ${color.yellow(projectPath)} is not empty. Proceed?`,
+    }))
+  );
 };
 
 type InitTSorJSProjectArg = {
@@ -306,9 +285,13 @@ async function initTSorJSProject({
 }: InitTSorJSProjectArg): Promise<void> {
   const frontendCompiledAquaPath = await ensureFrontendCompiledAquaPath();
   const indexFilePath = getFrontendIndexTSorJSPath(isJS);
-  const packageJSONPath = getPackageJSONPath();
+  const packageJSONPath = getFrontendPackageJSONPath();
   const frontendSrcPath = getFrontendSrcPath();
-  fluenceConfig.relaysPath = relative(projectRootDir, frontendSrcPath);
+
+  addRelayPathEntryToConfig(
+    fluenceConfig,
+    relative(projectRootDir, frontendSrcPath),
+  );
 
   const relativeFrontendCompiledAquaPath = relative(
     projectRootDir,
@@ -320,25 +303,81 @@ async function initTSorJSProject({
 
   await Promise.all([
     fluenceConfig.$commit(),
-    writeFile(indexFilePath, getIndexJsContent(isJS), FS_OPTIONS),
-    writeFile(packageJSONPath, jsonStringify(getPackageJSON(isJS)), FS_OPTIONS),
+    writeFile(indexFilePath, getFrontendIndexJsContent(isJS), FS_OPTIONS),
+    writeFile(
+      packageJSONPath,
+      jsonStringify(getFrontendPackageJSON(isJS)),
+      FS_OPTIONS,
+    ),
     writeFile(getIndexHTMLPath(), getIndexHTMLContent(isJS), FS_OPTIONS),
     isJS
       ? Promise.resolve()
-      : writeFile(getTsConfigPath(), TS_CONFIG_CONTENT, FS_OPTIONS),
+      : writeFile(
+          getFrontendTsConfigPath(),
+          FRONTEND_TS_CONFIG_CONTENT,
+          FS_OPTIONS,
+        ),
 
-    (async () => {
-      return compileToFiles({
-        compileArgs: {
-          filePath: await ensureAquaMainPath(),
-          imports: await getAquaImports({
-            maybeFluenceConfig: fluenceConfig,
-          }),
-        },
-        targetType: isJS ? "js" : "ts",
-        outputPath: frontendCompiledAquaPath,
-      });
-    })(),
+    compileToFiles({
+      compileArgs: {
+        filePath: await ensureAquaMainPath(),
+        imports: await getAquaImports({
+          maybeFluenceConfig: fluenceConfig,
+        }),
+      },
+      targetType: isJS ? "js" : "ts",
+      outputPath: frontendCompiledAquaPath,
+    }),
+  ]);
+}
+
+type InitTSorJSGatewayProjectArg = {
+  isJS: boolean;
+  fluenceConfig: FluenceConfig;
+};
+
+async function initTSorJSGatewayProject({
+  isJS,
+  fluenceConfig,
+}: InitTSorJSGatewayProjectArg): Promise<void> {
+  const gatewayCompiledAquaPath = await ensureGatewayCompiledAquaPath();
+  const indexFilePath = getGatewayServerTSorJSPath(isJS);
+  const packageJSONPath = getGatewayPackageJSONPath();
+  const gatewaySrcPath = getGatewaySrcPath();
+  const gatewayReadmePath = join(getGatewayPath(), README_MD_FILE_NAME);
+
+  addRelayPathEntryToConfig(
+    fluenceConfig,
+    relative(projectRootDir, gatewaySrcPath),
+  );
+
+  await Promise.all([
+    fluenceConfig.$commit(),
+    writeFile(indexFilePath, getGatewayIndexJsContent(isJS), FS_OPTIONS),
+    writeFile(
+      packageJSONPath,
+      jsonStringify(getGatewayPackageJSON(isJS)),
+      FS_OPTIONS,
+    ),
+    writeFile(gatewayReadmePath, GATEWAY_README_CONTENT, FS_OPTIONS),
+    isJS
+      ? Promise.resolve()
+      : writeFile(
+          getGatewayTsConfigPath(),
+          GATEWAY_TS_CONFIG_CONTENT,
+          FS_OPTIONS,
+        ),
+
+    compileToFiles({
+      compileArgs: {
+        filePath: await ensureAquaMainPath(),
+        imports: await getAquaImports({
+          maybeFluenceConfig: fluenceConfig,
+        }),
+      },
+      targetType: isJS ? "js" : "ts",
+      outputPath: gatewayCompiledAquaPath,
+    }),
   ]);
 }
 
@@ -365,7 +404,7 @@ function getIndexHTMLContent(isJS: boolean) {
 `;
 }
 
-function getIndexJsContent(isJS: boolean) {
+function getFrontendIndexJsContent(isJS: boolean) {
   return `import { Fluence } from "@fluencelabs/js-client";
 import relays from "./relays.json" assert { type: "json" };
 import {
@@ -479,7 +518,147 @@ function stringifyError(e${isJS ? "" : ": unknown"}) {
 `;
 }
 
-const TS_CONFIG_CONTENT = `{
+function getGatewayIndexJsContent(isJS: boolean) {
+  return `${
+    isJS
+      ? ""
+      : 'import type { TypeBoxTypeProvider } from "@fastify/type-provider-typebox";\n'
+  }import { Fluence } from "@fluencelabs/js-client";
+import relays from "./relays.json" assert { type: "json" };
+import { ${isJS ? "" : "type Static, "}Type } from "@sinclair/typebox";
+import fastify from "fastify";
+
+import { helloWorld, helloWorldRemote, showSubnet, runDeployedServices } from "./compiled-aqua/main.js";
+
+// This is an authorization token for the gateway service.
+// Remember to generate the appropriate token and save it in env variables.
+const ACCESS_TOKEN = "abcdefhi";
+
+// This is the peer's private key.
+// It must be regenerated and properly hidden otherwise one could steal it and pretend to be this gateway.
+const PEER_PRIVATE_KEY = new TextEncoder().encode(
+  new Array(32).fill("a").join(""),
+);
+
+const server = fastify({
+  logger: true,
+})${isJS ? "" : ".withTypeProvider<TypeBoxTypeProvider>()"};
+
+await server.register(import("@fastify/rate-limit"), {
+  max: 100,
+  timeWindow: "1 minute",
+});
+
+server.addHook("onReady", async () => {
+  await Fluence.connect(relays[0], {
+    keyPair: {
+      type: "Ed25519",
+      source: PEER_PRIVATE_KEY,
+    },
+  });
+});
+
+server.addHook("onRequest", async (request, reply) => {
+  if (request.headers.access_token !== ACCESS_TOKEN) {
+    await reply.status(403).send({
+      error: "Unauthorized",
+      statusCode: 403,
+    });
+  }
+});
+
+server.addHook("onClose", async () => {
+  await Fluence.disconnect();
+});
+
+const callbackBody = Type.Object({
+  name: Type.String(),
+});
+
+${isJS ? "" : "type callbackBodyType = Static<typeof callbackBody>;"}
+
+const callbackResponse = Type.String();
+
+${isJS ? "" : "type callbackResponseType = Static<typeof callbackResponse>;"}
+
+const showSubnetResponse = Type.Array(
+  Type.Object({
+    host_id: Type.Union([Type.String(), Type.Null()]),
+    services: Type.Union([Type.Array(Type.String()), Type.Null()]),
+    spells: Type.Union([Type.Array(Type.String()), Type.Null()]),
+    worker_id: Type.Union([Type.String(), Type.Null()]),
+  }),
+);
+
+${
+  isJS ? "" : "type showSubnetResponseType = Static<typeof showSubnetResponse>;"
+}
+
+const runDeployedServicesResponse = Type.Array(
+  Type.Object({
+    answer: Type.Union([Type.String(), Type.Null()]),
+    worker: Type.Object({
+      host_id: Type.String(),
+      pat_id: Type.String(),
+      worker_id: Type.Union([Type.String(), Type.Null()]),
+    }),
+  }),
+);
+
+${
+  isJS
+    ? ""
+    : "type runDeployedServicesResponseType = Static<typeof runDeployedServicesResponse>;"
+}
+
+// Request and response
+server.post${
+    isJS ? "" : "<{ Body: callbackBodyType; Reply: callbackResponseType }>"
+  }(
+  "/my/callback/hello",
+  { schema: { body: callbackBody, response: { 200: callbackResponse } } },
+  async (request, reply) => {
+    const { name } = request.body;
+    const result = await helloWorld(name);
+    return reply.send(result);
+  },
+);
+
+// Fire and forget
+server.post("/my/webhook/hello", async (_request, reply) => {
+  void helloWorldRemote("Fluence");
+  return reply.send();
+});
+
+server.post${isJS ? "" : "<{ Reply: showSubnetResponseType }>"}(
+  "/my/callback/showSubnet",
+  { schema: { response: { 200: showSubnetResponse } } },
+  async (_request, reply) => {
+    const result = await showSubnet();
+    return reply.send(result);
+  },
+);
+
+server.post${isJS ? "" : "<{ Reply: runDeployedServicesResponseType }>"}(
+  "/my/callback/runDeployedServices",
+  { schema: { response: { 200: runDeployedServicesResponse } } },
+  async (_request, reply) => {
+    const result = await runDeployedServices();
+    return reply.send(result);
+  },
+);
+
+server.listen({ port: 8080, host: "0.0.0.0" }, (err, address) => {
+  if (err !== null) {
+    console.error(err);
+    process.exit(1);
+  }
+
+  console.log(\`Server listening at \${address}\`);
+});`;
+}
+
+const FRONTEND_TS_CONFIG_CONTENT = `{
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
@@ -504,7 +683,29 @@ const TS_CONFIG_CONTENT = `{
 }
 `;
 
-function getPackageJSON(isJS: boolean) {
+const GATEWAY_TS_CONFIG_CONTENT = `{
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "skipLibCheck": true,
+    "resolveJsonModule": true,
+    "outDir": "./dist",
+    "verbatimModuleSyntax": true,
+    "esModuleInterop": true,
+
+    /* Linting */
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noFallthroughCasesInSwitch": true,
+    "forceConsistentCasingInFileNames": true
+  },
+  "include": ["src"]
+}
+`;
+
+function getFrontendPackageJSON(isJS: boolean) {
   return {
     private: true,
     version: "0.0.0",
@@ -520,7 +721,86 @@ function getPackageJSON(isJS: boolean) {
     },
     devDependencies: {
       vite: "4.4.5",
-      ...(isJS ? {} : { typescript: "5.0.2" }),
+      ...(isJS
+        ? {}
+        : {
+            [TYPESCRIPT_NPM_DEPENDENCY]:
+              CLIPackageJSON.devDependencies[TYPESCRIPT_NPM_DEPENDENCY],
+          }),
     },
   };
+}
+
+function getGatewayPackageJSON(isJS: boolean) {
+  return {
+    private: true,
+    name: "gateway",
+    description: "Fluence gateway for running aqua function via HTTP methods",
+    version: "0.0.1",
+    type: "module",
+    keywords: [],
+    scripts: {
+      ...(isJS
+        ? {}
+        : {
+            build: "rm -rf ./dist && tsc -p tsconfig.json",
+          }),
+      start: isJS
+        ? "node --no-warnings src/server.js"
+        : "node --no-warnings --loader ts-node/esm src/server.ts",
+    },
+    dependencies: {
+      [JS_CLIENT_NPM_DEPENDENCY]:
+        CLIPackageJSON.dependencies[JS_CLIENT_NPM_DEPENDENCY],
+      fastify: "4.25.2",
+      "@fastify/rate-limit": "9.1.0",
+      "@sinclair/typebox": "0.32.11",
+    },
+    ...(isJS
+      ? {}
+      : {
+          devDependencies: {
+            "@types/node": "20.11.5",
+            "@fastify/type-provider-typebox": "4.0.0",
+            [TYPESCRIPT_NPM_DEPENDENCY]:
+              CLIPackageJSON.devDependencies[TYPESCRIPT_NPM_DEPENDENCY],
+            "ts-node": "10.9.2",
+          },
+        }),
+  };
+}
+
+const GATEWAY_README_CONTENT = `## Fluence HTTP Gateway
+
+Here you can find an example of simple Fluence HTTP gateway.
+The gateway allows you to map Aqua functions to http requests i.e., you can create HTTP route and handle the incoming request by executing some Aqua function.
+You can check it out and test this repo.
+
+### Start gateway
+
+- \`npm install\`
+- \`npm run start\`
+- \`curl -X POST http://localhost:8080/my/callback/hello -H "ACCESS_TOKEN: abcdefhi" -H 'Content-Type: application/json' -d '{"name": "Fluence" }'\`
+- After running these commands you should see: \`Hello, Fluence\`
+
+### Notes:
+
+Gateway contains four routes corresponding to Aqua functions: \`helloWorld\`, \`helloWorldRemote\`, \`showSubnet\` and \`runDeployedServices\`.
+
+You can run \`helloWorld\` and \`helloWorldRemote\` right away.
+To run \`showSubnet\` and \`runDeployedServices\` successfully, it is required to do \`fluence deal deploy\` then compile Aqua and restart the server.
+
+Currently, \`fluence aqua\` won't put generated files to gateway dir automatically. Remember to pass \`-o/--output\` flag to CLI for that or change the config in \`fluence.yaml\`.
+
+> Remember to replace hardcoded token and peer private key. You can achieve that by placing these credentials in env variables, for example.
+`;
+
+function addRelayPathEntryToConfig(fluenceConfig: FluenceConfig, path: string) {
+  if (fluenceConfig.relaysPath === undefined) {
+    fluenceConfig.relaysPath = [path];
+  } else if (typeof fluenceConfig.relaysPath === "string") {
+    fluenceConfig.relaysPath = [fluenceConfig.relaysPath, path];
+  } else {
+    fluenceConfig.relaysPath = [...fluenceConfig.relaysPath, path];
+  }
 }