+++ src/lib/multiaddres.ts
@@ -41,7 +41,11 @@ import {
   CONTRACTS_ENV,
 } from "./const.js";
 import type { ProviderConfigArgs } from "./generateUserProviderConfig.js";
-import { commaSepStrToArr, jsonStringify } from "./helpers/utils.js";
+import {
+  commaSepStrToArr,
+  jsonStringify,
+  splitErrorsAndResults,
+} from "./helpers/utils.js";
 import { projectRootDir } from "./paths.js";
 import { input, list } from "./prompt.js";
 import { resolveFluenceEnv } from "./resolveFluenceEnv.js";
@@ -322,19 +326,16 @@ export async function updateRelaysJSON({
   noxes,
 }: UpdateRelaysJSONArgs) {
   if (
-    typeof fluenceConfig?.relaysPath !== "string" ||
+    fluenceConfig?.relaysPath === undefined ||
     envConfig?.fluenceEnv === undefined
   ) {
     return;
   }
 
-  if (isAbsolute(fluenceConfig.relaysPath)) {
-    commandObj.error(
-      `relaysPath must be relative to the root project directory. Found: ${
-        fluenceConfig.relaysPath
-      } at ${fluenceConfig.$getPath()}}`,
-    );
-  }
+  const relayPaths =
+    typeof fluenceConfig.relaysPath === "string"
+      ? [fluenceConfig.relaysPath]
+      : fluenceConfig.relaysPath;
 
   const relays = await resolveAddrsAndPeerIds({
     fluenceEnv: envConfig.fluenceEnv,
@@ -342,8 +343,29 @@ export async function updateRelaysJSON({
     noxes,
   });
 
-  await writeFile(
-    join(resolve(projectRootDir, fluenceConfig.relaysPath), "relays.json"),
-    jsonStringify(relays),
+  const [absolutePaths, relativePaths] = splitErrorsAndResults(
+    relayPaths,
+    (relayPath) => {
+      return isAbsolute(relayPath)
+        ? { error: relayPath }
+        : { result: relayPath };
+    },
+  );
+
+  if (absolutePaths.length > 0) {
+    commandObj.error(
+      `relaysPath must contain only paths relative to the root project directory. Found: ${absolutePaths.join(
+        "\n",
+      )}`,
+    );
+  }
+
+  await Promise.all(
+    relativePaths.map((relativePath) => {
+      return writeFile(
+        join(resolve(projectRootDir, relativePath), "relays.json"),
+        jsonStringify(relays),
+      );
+    }),
   );
 }